♻️ refactor: 도메인 패키지 구조 분리 (id/, enums/)
- OutboxId, NotificationId를 domain/id/ 패키지로 분리
- OutboxStatus enum을 domain/enums/ 패키지로 분리
- OutboxId에 0 이하 id 검증 추가
