+++ notification-core/src/main/java/com/liveklass/notification/domain/id/NotificationId.java
@@ -1,0 +1,8 @@
+package com.liveklass.notification.domain.id;
+
+public record NotificationId(Long id) {
+
+    public static NotificationId of(final Long id) {
+        return new NotificationId(id);
+    }
+}
